updated the test drivers to dealocate the memory

+++ newpolka/tests/libFuzzer/test_poly.h
@@ -14,14 +14,14 @@
 #define R(i) i < K
 
 bool create_polyhedron(pk_t** polyhedron, ap_manager_t* man, pk_t * top,
-		long dim, const long *data, size_t dataSize,
+		int dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp);
 
 bool make_fuzzable(void *array, size_t size, const long *data, size_t dataSize,
 		unsigned int *dataIndex);
 bool assume_fuzzable(bool condition);
 
-bool make_fuzzable_dimension(long * dim, const long *data, size_t dataSize,
+bool make_fuzzable_dimension(int * dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp);
 
 #endif /* TEST_POLY_H_ */
